Showreel specific Dynamic Text component
Summary:
This bridges a dynamic text component for showreel. It's loosely based off of AOSP `autoSizeText` which can be found here: https://fburl.com/dn7vzghv

The TLDR is that we do a binary search on an ordered list of text sizes and finds the best one that fits. To do that we create a static layout using the text paint. The search criteria is if static layout height is greater than the heightSpec height passed to the dynamic text during layout.

If it does not fit within the minimum text size, we use the minimum size.

We purposefully ignore any text size set on text spans.

**This diff bridges behaviour for the following props:**
* children
* max_text_size
* min_text_size
* text_align

The following diffs in the stack will focus on other props to help diff readability.
**I will squash all these diffs into one before landing**

Reviewed By: iamsadesh

Differential Revision: D26891799

fbshipit-source-id: d64b517cbac44382fb2f055187b6030e329d22e6