Preload KComponent class earlier in FbFragmentActivity
Summary: This is to unblock Workplace release and verify that the issue they are hitting in T90677996 is caused by a race condition between running the Litho startup logic and rendering initial feed. If this fixes their issue, we can work on a more correct fix that doesn't need to be merged into a release branch.

Reviewed By: dezfx

Differential Revision: D28413552

fbshipit-source-id: 88cae33dbd8942bca6cfe47c4013b09d7d1df9b9