Rebind triggers under lock
Summary: Since we clear triggers and then re-add them, if a trigger is sent from the main thread which this is happening in the background, it could cause the trigger not to be found even though it hasn't gone away. Rebinding under the same lock we use to look up triggers should help.

Reviewed By: adityasharat

Differential Revision: D28540943

fbshipit-source-id: d748e1792a981f487e837acf85a363cdc9470335