Support StartSnapHelper scrolling for grids
Summary:
As a context, I needed the SNAP_TO_START effect for grids on a feature that Messenger Kids needs. But while implementing as the usual, the scrolling was not as expected.
See the video showing the buggy scrolling for grids below. As we can see in the video, when we scroll from column 1 to column 2, it's like the scrolling fails to finish its action and column 2 doesn't take column 1 position but comes back to its original position before the scroll.

The method findTargetSnapPosition() is returning the wrong position to snap to for grids. When we snap forward from column 1 to column 2, findTargetSnapPosition() returns the position of the second item in column 1 like a list would do it. That means findFirstViewBeforeStart() returns exactly the view before the snap then add the flingOffset to find the next view to focus on. This is the right way for list.
But for grids, we need to calculate based on the number of columns/rows and the flingOffset.

Reviewed By: astreet

Differential Revision: D28558384

fbshipit-source-id: e294bd5882648d8bf51cdbf6ee7350ac069129c3