Ensure moveToLifecycle is called on UI thread
Summary:
Enforce  `LithoLifecycleProviderDelegate.java#moveToLifecycle` to be called on the UI thread, since it performs actions such as releasing the CT or setting visibility hint which are also UI thread only.
This also ensures that `moveToLifecycle` is performed on a single thread atomically and there can't be bugs caused by interleaving state changes, which can lead to bugs such as in t89281152.
An example of the bug that this fixes can be seen in ComponentTreeHolderLifecycleProvider. What can happen:
- request moveToLifecycle(DESTROYED) and change state to DESTROYED
- Because `moveToLifecycle` is not an atomic operation, before executing `mParentLifecycle.removeListener(this)` and removing the destroyed `ComponentTreeHolderLifecycleProvider` as a listener, receive moveToLifecycle(HINT_VISIBLE) notification from lifecycle provider. Attempt to change state from DESTROYED to HINT_VISIBLE which is invalid

Reviewed By: adityasharat

Differential Revision: D28599815

fbshipit-source-id: 9b90f1c45d0f2c1a3195607cddbfb6ebd075833e