Sync Return Type of Service-Aware Methods to OnCreateService
Summary:
Service-aware methods can contain service parameter which is a return type of `OnCreateService`. This diff assumes the above sentence and parses the service type from it.

To do that, `SpecModel` for section spec needs to be added. The test for it also added here. The `SpecModel` was partially copied from D6591949.

This will be landed after the below diffs are published as beta version of Litho plugin.

Reviewed By: Katalune

Differential Revision: D27291143

fbshipit-source-id: 16f7e96087285d2f3e236478ef81fe2813937548