Passes LayoutStateContext to InternalNode#freeze API
Summary:
**API only changes**

When `InternalNode` gets reused the layout results need to use the latest `ComponentContext` and therefore MUST not access context from the `InternalNode`.

This diff is the first step to make those changes.

This diff passes the `LayoutStateContext` via the `freeze` API. This `LayoutStateContext` will be used to get the latest `ComponentContext` at layout time.

Since the `YogaNode.Input.freeze` API cannot be used for this purpose the `InputOnlyInternalNode` creates a container object to hold the `LayoutStateContext` and the `InternalNode` to delegate the call as follows:

```
YogaNode -> LayoutContextContainer.freeze(node, parent) -> InternalNode.freeze(c, node, parent)
```

`DefaultInternalNode` can continue to use its `mComponentContext` field.

Reviewed By: mihaelao

Differential Revision: D27905578

fbshipit-source-id: d2837b87816318ba0f95fa464aa85259419d6f6c