DynamicText: Support background color
Summary:
Reuse `RoundedCornerTextBackgroundSpan` to draw background span.

Since the background span introduces "padding", we need to account for it when calling `autoSizeText`. We adjust the width and height spec when calculating the text size that fits.

To account for clipping left and right:
We need to adjust the draw after to shift by the background padding by introducing `extraSpacingLeft` and `extraSpacingRight`.

To account for top bottom clipping:
We set the vertical gravity to center.

Edge case: When text doesn't fit, we use a hardcoded padding and radius. Otherwise the it would not align correctly.

Reviewed By: pasqualeanatriello

Differential Revision: D28085710

fbshipit-source-id: 9fabc327a03c139b1cf9d3a4cb42ade3fb97bfbe