Don't destroy nested ComponentTree in Vertical/Horizontal scrolls
Summary:
HorizontalScrollSpec and VerticalScrollSpec both keep a nested ComponentTree as a StateValue.
When a lifecycle provider is hooked to them, as long as the ComponentTree is a StateValue we can't move the nested tree to a destroyed state.
That's because when used as an item in a RecyclerBinder, we can be destroying the parent ComponentTree when moving outside the prepared window, but keep the StateHandler (containing the ComponentTree) in case the item goes back into the prepared window. if we do that, we'd be trying to use the nested ComponentTree after it was released, leading to crashes like T92809531

Reviewed By: adityasharat

Differential Revision: D29099452

fbshipit-source-id: c666217c5a39df28e36f5aeffd81d01ea355f0b8