Update MountStateViewClickTest to use a MountSpec instead of TestViewComponent
Summary:
The 1st two tests in MountStateViewClickTest.java were using TestViewComponent which is deprecated. Instead, a new LayoutSpec was created which holds a simple mount spec that can be assigned with optional click and long click handlers.

Furthermore, the relevant tests no longer call "mountComponents", but instead, use the LithoViewRule to replace this process. To achieve this, a helper method was added to properly set the LithoView's parent and component tree (similar to what happens in mountComponents). An additional test in this file that also used a similar flow was adapted to reuse this helper method.

Finally, the two relevant tests were expanded to ensure long click is disabled when click is expected and vise-versa.

Reviewed By: adityasharat

Differential Revision: D29129196

fbshipit-source-id: e7bd75af6453e396b5aae2fc293e006e8ac1819d