Remove redundant test from TreeDiffingTest.java
Summary: TreeDiffingTest.java had a "white-box" test called testIsEquivalentToCalled which was checking if "isEquivalent" was being called during transitioning from similar layouts. This test was deemed too "microscopic", and it was decided that a more general test of the public API was required. Upon further examination, it was noted that such a test already exists in LayoutDiffingTest.java: "onSetRootWithSimilarComponent_thenShouldNotRemeasureMountSpec". As such, this "isEquivalent" test and related code was removed.

Reviewed By: adityasharat

Differential Revision: D29133332

fbshipit-source-id: 0437c045aa2c587034503492d6a32cc2d015c253