Fixed Component tree bug for LithoViewRule
Summary: When using a new lithoview on the LithoViewRule the LithoViewRule would automatically override the component tree with whatever it already had if its not null and not use the component tree tied to the lithoview, causing problems. This diff checks if the component tree on the new lithoView, to see if it is null, if not it will use this tree.

Reviewed By: adityasharat

Differential Revision: D29160662

fbshipit-source-id: aea283bd4004444502a5f19096a23bc2eb4122b3