Introduce a StyleCompat functions to pass KComponent styles from Java code
Summary:
See next diff on how this will be used. I have not filled all the details necessary.
This creates a StyleCompat scheme to allow building a style object in Java call sites. It's going to be a lot of functions to add and make it work.
Note the `when`s which will be here to keep the style consistent for Java.

Reviewed By: astreet

Differential Revision: D29175938

fbshipit-source-id: d3d23c6f8f01d75c42a4328d4c23b337e4c9e7a1