Solve root non-mount-view spec's view attributes from being applied on LithoView when shouldDisableDrawableOutputs is true
Summary:
When setting shouldDisableDrawableOutputs to true, Litho will avoid creating drawables for backgrounds and foregrounds. However, in this configuration, there is an edge case where a non-view-mount-spec's view attributes (e.g., bg, fg, click handler, etc) are set on the root LithoView. This is problematic, as the LithoView and the relevant spec may not have the same bounds, and the result would apply such attributes on the whole LithoView rather than the intended nested spec.

To solve this, when calculating and setting a host view in these circumstances, root views will get their own host rather then that host becoming the LithoView. In addition, needsHostView will now ignore the is-root-layout condition when the flag is on.

Reviewed By: adityasharat

Differential Revision: D29159472

fbshipit-source-id: 6452f6f3f0945fb706a09f44b26d403defebacf9