Created Test to check ComponentTree resets visibility flags on detached LithoView
Summary:
A ComponentTree that listens to a LithoLifecycleProvider and is attached to a LithoView will set the visibility hint flags on LithoView accordingly.
When the LithoView is detached from the ComponentTree, the visibility hint flags on the LithoView should be reset - otherwise it may be stuck in a permanent "invisible" status.

This diff creates a test to check that the visibility hint on the LithoView is reset as fixed in D29299640 (https://github.com/facebook/litho/commit/206ecd82dc8950b44495776b29760fd0a0a86e9c)

Reviewed By: mihaelao

Differential Revision: D29457440

fbshipit-source-id: 1682c3651f66945821631a1df0d9b821f9ff4ecd