Fix generation of getter for boolean in kotlin
Summary: strulovich highlighted that for java specs written in kotlin with prop default we try to access the value using a getter. This diff changes the generator code to access values properly with getter only if they don't start with "is". (We could assume that those will be only boolean values and check on boolean as well, but then we would add additional check, and it seems that the TypeSpec cannot be checked based on the TypeName as we can only check on DeclaredTypeSpec : https://fburl.com/code/ewgy780z and we would have to check against the String value)

Reviewed By: strulovich

Differential Revision: D29392778

fbshipit-source-id: e2f6abb06e61424547d5dfd8230fd9de2f75a303