Add tests for capturing props/state from Collection component
Summary:
These tests just verify existing behavior. As mentioned in two of these tests, the behavior isn't desired, but we should encode it as part of our contract until we change it, so that we don't change it accidentally.

This diff is also in preparation for introducing a utility to solve this problem.

Reviewed By: strulovich, colriot

Differential Revision: D29431945

fbshipit-source-id: bcb98c6817834bf98f31bdf1a24ab9ad6830e870