Fix memory leak on RCTextView
Summary: The issue seems to be that a reference to a context in this case `ReelViewerFragment` is being kept by an `RCTextView`. `RCTextView` are being pooled and `mount` and `unmount` is what we use to get them ready to be reused. On the bloks side when Clickable spans are created (if the text view needs to create them ofc) https://fburl.com/code/amfijkux they hold a reference to the context. I think the issue is that this clickable spans get assigned to `RCTextView` on `mount` https://fburl.com/code/v81tv1gz, but then on `unmount` they don't get cleared https://fburl.com/code/gkmjtww6. Because of this we might have this views pooled, holding references to contexts that are not supposed to exist anymore.

Reviewed By: nicous

Differential Revision: D29681349

fbshipit-source-id: db7e8d42fa6e30480a3a799f586309e256332340