InternalNode and LithoLayout extends RC Node and LayoutResult
Summary:
InternalNode and LithoLayout extends RC Node and LayoutResult.

The API `InternalNode#calculateLayout()` needs to pass in the current `LithoLayoutResult` to implement caching. Since RC's `LayoutResult` API is purpose built for this using the `LayoutContext` and `RenderContext` generic, this diff essentially makes `LithoLayoutResult` implement the `LayoutResult` interface; which is evident what it will be eventually.

Reviewed By: mihaelao

Differential Revision: D29515792

fbshipit-source-id: c51aaff98c5f51b8d36935338fe55ea1212ebd43