Convert classes to Kotlin
Summary:
Continuation of the previous diff.
astreet suggested to convert stuff to kotlin and it turned out to be quite a refactor so im putting new diff with the changes to files and acitvities and leaving the previous diff only about moving kotlin sample app into java.

I added a way to show component by diretly passing it to the demo instead of using component creator (we dont need a context for kotlin components)

Reviewed By: astreet

Differential Revision: D29587804

fbshipit-source-id: 8e9e8306924e6b3125ed8c36dbe1f64fbfaa8cf1