StateContainers are added to the StateHandler during reconciliation
Summary:
When a component's internal state is not updated, the current implementation is also marks their StateContainer as being used, so that they aren't pruned when the StateHandler is committed.

This diff ensures the same when InternalNode Nodes are reused or reconciled, in which case the internal state is not updated for their components.

Reviewed By: mihaelao

Differential Revision: D29710124

fbshipit-source-id: 2ff541a92d5d59570902cae6d61ce8e6e15a4e80