Add a reference to the associated EditText view in EditorActionEvent (#807)
Summary:
Add a reference in EditorActionEvent to the EditText view associated with the event. This allows the users to access the underlying EditText view when the event is received, similarly to the TextChangedEvent.

## Changelog

Add a reference to the associated EditText view in EditorActionEvent.

Pull Request resolved: https://github.com/facebook/litho/pull/807

Test Plan: No relevant tests; I confirmed that `dispatchEditorActionEvent` is not called in any other places.

Reviewed By: adityasharat

Differential Revision: D29760914

Pulled By: astreet

fbshipit-source-id: 6ad5319d60548545a6aee4cf933a219c35547ae0