Fix crash in TransitionsExtension due to parent being unmounted
Summary:
In certain edge cases, the TransitionsExtension will attempt to remount unmounted disappearing items and this will result in a crash due to the fact that the disappearing item's parent has already been unmounted. This fix aims to solve this issue by scanning the parents of the disappearing items and ensuring they are mounted as well.
To achieve this, the code has been slightly refactored to ensure all parents are mounted before beginning to process the disappearing child and its descendants. Some additional refactors were made for code clarity.

Reviewed By: nicous

Differential Revision: D29762464

fbshipit-source-id: 8a0f4ca82fd6f649a48f882ebc9b0a50af4d1a48