Workaround for launch blocker T96573697: Create local copy of mLithoView reference to avoid unexplained null
Summary:
mLithoView is becoming null at some point during the execution of ComponentTree.mountComponentInternal. We do not know what is causing this, although it's likely a race condition.
Creating a local copy of the mLithoView reference so that it is guaranteed to not be null. Additional logging added to track this issue.

Reviewed By: mihaelao

Differential Revision: D29939882

fbshipit-source-id: d93098f2af32e750883ee2ce32b24967366d51bc