Apply value for shouldDisableBgFgOutputs when setting value on shouldAddHostViewForRootComponent in TempConfigurations, and fix test that used this configuration incorrectly.
Summary:
In the previous setup, when manually setting true in ComponentConfigurations on the relevant configs, several tests would fail. This fix ensures that both values are correctly set and restored during test runtime.
In addition, LayoutStateCalculationWithoutDrawableOutputsTest was incorrectly adjusting the value of this configuration and needlessly caching a value. Adapted this test to use TempComponentsConfigurations

Reviewed By: adityasharat

Differential Revision: D29934861

fbshipit-source-id: 0035be61a4efd02b230d2cd579bafdd7aa07e91a