Removes OnPrepare benchmark metric
Summary: Removes OnPrepare benchmark metric since the OnPrepare phase cannot be run independently of the Layout creation phase. The current implementation relies on a hack solution to initial the internal state of the component artificially. This hack cannot be support with stateless components so this diff removes this benchmark metric.

Reviewed By: SidharthGuglani

Differential Revision: D29991427

fbshipit-source-id: ce91705b9b322d356adb7f071bbda94cb954fbd9