Fix crash in AOSPLithoLifecycleProvider
Summary:
When we add observer to Android Lifecycle it's tries to calculate the target state of observer which is passed as param. After calculating the target state, it dispatch events to till it reaches it's target state. (AOSP code link https://fburl.com/7fi2vlyu)

In constructor we were adding the observer before initialising the `mLithoLifecycleProviderDelegate` leading to NPE.

Reviewed By: adityasharat

Differential Revision: D30039127

fbshipit-source-id: fd94823b7ad4c878a88a93d694b242e33965708c