Check common props in isComponentEquivalent
Summary:
The kotlin builder for `Text` accepts style values, some of which are applied directly as common props. Common props are not currently compared in our `onCheckIsSameContentEventHandler` so state updates are missed.

e.g. In the following, only the initial value of `alpha` will be applied.

```
val alpha = useState { 1f }

Collection {
  staticChild { Text("Text", style = Style.alpha(alpha.value) }
}
```

Fixing this by also checking commonProps in `onCheckIsSameContentEventHandler`

Reviewed By: astreet

Differential Revision: D30105106

fbshipit-source-id: afb3a58624d10772ab4debc196063426fd8b7d8d