Uses the index in the parent to mount and unmount
Summary:
Uses the index in the parent to mount and unmount. Right now all the indexes which were used to mount and unmount were the absolute index of the item in mountable output list.

This diff changes that to use the index in the parent host. This will allow removing the index in the LayoutOutput and it should also make mount/unmount slightly more optimised for now ComponentHost will use smaller sparse arrays.

Reviewed By: nicous

Differential Revision: D30192466

fbshipit-source-id: c78a52fa64b05c9d54e76e181d782b946e032aeb