Add a rawInputType parameter to TextInputSpec (#810)
Summary:
Add a rawInputType parameter to TextInputSpec. Being able to set rawInputType allows user to achieve a multiline EditText instance with a Done button. The handling of the parameter, i.e. not setting inputType if rawInputType is specified, as well as the param description, is the same as in the old EditTextSpec (litho/litho-widget/src/main/java/com/facebook/litho/widget/EditTextSpec.java).

## Changelog

Add a rawInputType parameter to TextInputSpec.

Pull Request resolved: https://github.com/facebook/litho/pull/810

Test Plan: N/A. Android doesn't expose a getRawInputType method, only setRawInputType. The field has also been added to MaterialTextInputSpec, and I've verified that the methods with modified signatures are not used anywhere else in Litho.

Reviewed By: colriot

Differential Revision: D30366802

Pulled By: astreet

fbshipit-source-id: a3b1d7e5d66934c70d73ecb01d8ee3ff7120853b