Add ability for all Litho unit tests to run under different Litho configurations
Summary:
This change makes it so all tests using LithoTestRunner run under default configuration as well as any other configurations specified in getExtraRunConfigurations. For instance, in this diff we add a StatelessComponents run configuration.

The purpose is so that as we iterate on framework changes, we make sure our unit test suite is passing for configs we plan to test in production automatically.

Reviewed By: adityasharat

Differential Revision: D30511504

fbshipit-source-id: c65a044a1e5100df095922a1b578d87986f5999e