Ensure host components do not get recycled in RenderCore MountState
Summary:
Similar to Litho's behaviour, we are now ensuring that hosts do not get recycled via the MountItemsPool in RenderCore MountState. This is done by adding an overridable method in RenderUnit "shouldAvoidRecycling", and in the LithoRenderUnit, overriding this method to return true when the component is a host component.

Also - cleaned up disableComponentHostPool ComponentsConfiguration flag, as the value is always true. Removed references in the code-base and defaulted behaviour to assume true.

Reviewed By: mihaelao

Differential Revision: D30766340

fbshipit-source-id: 16e4d1bee57c7be7661c2fdf8c24d578716a6213