Add equls/hashCode to Style
Summary:
Allow Styles to be compared. Have added tests as the equals check recurses along the linked list, which could be fragile if the data model changes.

This is necessary for checking equivalence of components that accept a `Style` as a prop, as prior to this change they will never be equivalent.
This will prevent unnecessary re-draws in Collections, which uses component equivalence to check if items have changed.

Differential Revision: D30775490

fbshipit-source-id: a146547257b09c51ae2f95cc642092dfb22d8b64