Remount host to root at the end of the root's mount items
Summary:
When remounting host to root in TransitionsExtension, we are incorrectly passing the index in the tree where the disappearing item was found as the index in the root where it should be added. Instead, since this is a disappearing item, it should always be added to the end.

Note: when adding this item to the root in this arbitrary position, it exposes a huge risk where another item can be moved into that position, leaving this item in a scrap list in the host, which will result in index-out-of-bounds exception when updating the child draw order in the host.

Reviewed By: adityasharat

Differential Revision: D30867987

fbshipit-source-id: dd68306420840edae5611027e93b409e663f8130