Stops the usage of bounds from LayoutOutput [3/n]
Summary:
Instead of using bounds in should update, explicitly use the width and height of the output instead.

This diff creates `LithoLayoutData` to store the width and height necessary for the should update of Litho Components  where `isMountSizeDependent` is true. This object will eventually be used to host the measure time Inter Stage Props as well.

Differential Revision: D30813166

fbshipit-source-id: b2cf693734037794a4be604dacb256d564a60be0