Ensure child draw order is dirty after unmounting a view
Summary:
In android, when a view is removed, it triggers a sequence that will check the child draw order before the view is actually removed from the viewgroup's child list. This means that when we call unmountView (which sets mIsChildDrawingOrderDirty to true at the beginning), getChildDrawingOrder is called within the view and the child drawing order is immediately marked as not dirty anymore while the view has not yet actually been removed. When the child is removed afterwards, the child drawing order array will still include indexes that assume that view is still there.

To solve this, after the view is removed from the parent, we reset the dirty flag to true again, so when the child order is queried again, it is recalculated without the removed view.

Reviewed By: adityasharat

Differential Revision: D30900331

fbshipit-source-id: ab304a5d129fc598726cb414db8424aa1753ecc8