Ensure shouldUpdate works correctly in RenderCore MountState
Summary:
In RenderCore MountState, updateMountItemIfNeeded utilizes mount extensions which (via binders) check to see if certain elements need to be unmounted or mounted. To do this, the flow checks if the LayoutState ids from both old and new components is different. If they are the same, an optimized check is performed that also ensures that both components are DrawableComponents. If either of them is not a DrawableComponent, then shouldUpdate returns false.

In Litho's MountState, this was done by checking the LayoutState instance passed into the mount method directly. However, in RenderCore's MountState, this LayoutState instance is only accessible via the context. However, by the time mount is called, this reference to the LayoutState has been nullifed. This means that the flow reaches this point, both LayoutState ids resolved to 0, falsely indicating that they are the same, and then always return false for shouldUpdate if either component isn't a drawable-component.

To correct this, now the current and previous layoutstate Ids are being passed thru LithoLayoutData, and the comparison is happening there.

Reviewed By: adityasharat

Differential Revision: D30931960

fbshipit-source-id: d28ad4c7a5175b182d8d19ce85ae8a6efa14d680