Sets the ComponentContext from the tail component during reconciliation
Summary:
The `ComponentContext` of a `LayoutResult` must be the from its `Component`. This diff fixes the bug where the `ComponentContext` from the head component is set instead.

The components of the working ranges do not need to be updated, because they already exist in the updated components list, and therefore can be read from it.

Reviewed By: SidharthGuglani

Differential Revision: D30993610

fbshipit-source-id: af37cc37e197a808dcc03eb26ae0e6228804d5bb