Ensure TransitionsExtension remounts at end of new layout root host
Summary:
Previously, TransitionsExtension was calculating the end of the root host by checking getMountItemCount of the outgoing frame. This was incomplete because we should consider both the incoming and outgoing frame. If the incoming frame has more children than the outgoing frame, the disappearing item must be added at the end of those to ensure it doesn't get overrideen when the incoming frame's children are added. We must still consider the outgoing frame's children, because if it has more children than the incoming frame, we don't want the disappearing item unmounted when the incoming frame cleans up the extra indices.

To solve this, we now use both the new incoming layout state from the extension input and the child count in the old frame (whichever is greatest). This will ensure that disappearing items are added at the very end and cannot be unmounted prematurely or overridden.

Reviewed By: astreet

Differential Revision: D30992098

fbshipit-source-id: 5374cb5df4ef92dd36cf40173685f2fa061aea39