Generate RenderTree prior to beforeMount() call
Summary: In TransitionsExtension.beforeMount, when delegating to rendercore mount, we now check the amount of children in the root, assuming the root will always be present. However, in LayoutState, there may be a case where there are 0 mountable outputs, in which case, the mountable-output list will be empty. Since RenderCore MountState assumes there will always be at least a root mountable output, the toRenderTree method artificially adds this if the list is empty. To ensure this behaviour exists during beforeMount, we now call toRenderTree prior to "beforeMount" from LithoView, rather than directly when mount is called.

Reviewed By: adityasharat

Differential Revision: D31053445

fbshipit-source-id: 59454dac6df796d04561cdb73aca08207868b42b