Fix Litho tests in MountStateTest failing when delegateToRenderCoreMount is true
Summary: The affected tests inside MountStateTest were force-casting the MountDelegateTarget to a Litho MountState, thereby resulting in a class-cast-exception. The only reason this was being done was to grab the DynamicPropsManager. To avoid this crash, added a VisibleForTesting method in LithoView that grabs the DynamicPropsManager appropriately, and added a way to get this manager from the DynamicPropsExtension. All added methods are VisibleForTesting only.

Reviewed By: adityasharat

Differential Revision: D30934708

fbshipit-source-id: bb16a09f41cb1c02151351951562bfe2412dc922