Kotlin state update test fixes, passing correct updated StateHandler to ComponentScope
Summary:
With stateless mode enabled, we do not create updated ComponentContext. KState still had access to older ComponentContext which does not have the latest StateHandler objects with updated states.

In this diff we are passing the StateHandler separately so that it can be used and we can have correct behaviour.

Reviewed By: adityasharat

Differential Revision: D31054005

fbshipit-source-id: aef0e5d05d7fa04fe2416630f43bbda8f6a01c59