Allow Setting Transition Key on Card Background
Summary:
**End goal:** Be able to animate the card background separately from the content within the card.  That way, the card content can show while adding an appear animation on the card background.

## Current Limitation
Currently, you can only set a transition on the card itself.  The issue here is that this means you're adding a transition for the card background *as well as the content of the card*.  You can't animate them separately, so you can't have the content show up and then the card background.

We'd like to be able to animate showing the card background separately from the content appear.  For the background, we want to use the scale animation, but if we currently do that, we'll end up scaling the entire card, including content.

## This Diff
Enables passing in a transition key that way the card background can be animated separately from the content within the card.

## Implementation Notes
In order to get this to work in the non-transparent card case, I needed to add a column that we can transition with the background set.  If there are other/better ways to handle the transparent card case, I'm all ears.

Reviewed By: mihaelao

Differential Revision: D31150883

fbshipit-source-id: 6f282141c50694856157559e2ecb1bb663e22867