Remove implementation of OnShouldCreateLayoutWithNewSizeSpec
Summary:
We're doing some modernization of Litho internals and noticed this feature is extremely rarely used so we'd like to remove it to simplify the library. This method is generally only called in the context of a "double measure" - this means in the same layout, we need to measure the same component twice e.g. to first determine how big it wants to be and then give it a different size.

We plan to provide new forms of caching component creation in the coming months.

Note: this does not yet remove the code from the AP, will do in follow up

Reviewed By: mihaelao

Differential Revision: D31277555

fbshipit-source-id: b2806fa652d1efce12b93b2b3847c5c7b656a886