Correct logic of LithoViewAttributesExtension's shouldUpdate
Summary:
In Litho's MountState, attributes are updated if the component should update, or the node-info should update. In LithoViewAttributesExtension, only the latter was being checked. To achieve parity with Litho's MountState, this logic has been corrected to properly OR the 2 conditions.
Note: Some code was move around in order to reuse the shouldUpdate logic in the LithoViewAttributesExtension.

Reviewed By: adityasharat

Differential Revision: D31382389

fbshipit-source-id: 07b7af767f887700f47d2e7121b3d6c8caa6fe1f