Remove ComponentTree when Vertical/Horizontal Spec unmounts
Summary:
This was already being done before D27618704 (https://github.com/facebook/litho/commit/dd0b82d3b29018701160076b127e6e957d3b4602), but was removed because we didn't unmount the content of the nested LithoView when the parent was being remounted.
However if we don't null out the ComponentTree on the LithoView, since the ComponentTree is kept as a state, if we remount the same layout we won't remount after unmounting because this check will fail https://fburl.com/code/imxuih5s

With this change we remove the ComponentTree when unmounting, but prevent unmounting its content to avoid bugs such as t88358575. Change only applies to Vertical/HorizontalScrollSpecs

Differential Revision: D31338456

fbshipit-source-id: 84eabdbd556ee4dea86314c0d3eab39c69cea827