Show tooltips on UFI Share button
Summary:
If eligible for showing, trigger a tooltip on the UFI share button.

IgdsTooltip modifies the container view hierarchy to show the tooltip view by adding the tooltop view as a child of the container. We can't have that container be a LithoView, since modifying the LithoView hierarchy manually is not allowed. This wraps the LithoView in a FrameLayout which we can use as container for the tooltip.

This diff triggers the tooltip to show, but I'll need some help figuring out how to position it correctly. The `LithoTooltip` API provides the LithoView reference and the bounds of the component we want to use as anchor relative to that LithoView. How can we position the IgdsTooltip correctly using this info?

Differential Revision: D31305936

fbshipit-source-id: 7128b68f64b7febf4417bbbbb8e21651f5252265