Set up an experiment for enforcing size 1 for all bg layout calculation thread pools
Summary:
Sets up an experiment to migrate all clients of ThreadPoolLayoutHandler to single threaded thread pools, after successful experiment and FAD improvements for News Tab. A few surfaces already want the thread pool to have 1 thread, so this will effectively be a no-op for them.

If we find this to be effective, we'll want to move from single threaded pools to just threads to simplify.

Reviewed By: mihaelao

Differential Revision: D31505419

fbshipit-source-id: d7f15fdc00f2097c15a3e40f00db04de1957ce3a