Added logging for releasing mount content that is still bound
Summary:
We are witnessing a crash where recycler-views are attempting to case their LayoutManager into the wrong type. The current theory is that this is only happening because somehow an item is being returned to teh mount item pool without being properly unbound. To test this theory, this diff adds an AuditableMountContent interface which LithoRecylerView implements, and checks the total amount of scroll listeners it has (as the crash is happening from an assigned scroll listener). If the scroll listener list is not empty, an exception will be thrown.

Note: RecyclerView does not expose its list of onScrollListeners, so to check that the list is not empty, we maintain a separate list of scroll listeners which is modified on any call to add, remove or clear.

Reviewed By: astreet

Differential Revision: D31545201

fbshipit-source-id: 572302e4b838364adc0c9d39bc805fcd194df462